requirements.txt
async-timeout==5.0.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect==1.2.0
Flask==3.1.0
Flask-Mail==0.10.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
pyquery==2.0.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
